+++ sigstore/verify/verifier.py
@@ -46,6 +46,7 @@ from sigstore._internal.sct import (
     _get_precertificate_signed_certificate_timestamps,
     verify_sct,
 )
+from sigstore._internal.timestamp import TimestampSource, TimestampVerificationResult
 from sigstore._internal.trust import ClientTrustConfig, KeyringPurpose, TrustedRoot
 from sigstore._utils import base64_encode_pem_cert, sha256_digest
 from sigstore.errors import VerificationError
@@ -58,6 +59,10 @@ _logger = logging.getLogger(__name__)
 # Limit the number of timestamps to prevent DoS
 # From https://github.com/sigstore/sigstore-go/blob/e92142f0734064ebf6001f188b7330a1212245fe/pkg/verify/tsa.go#L29
 MAX_ALLOWED_TIMESTAMP: int = 32
+
+# When verifying a timestamp, this threshold represents the minimum number of required
+# timestamps to consider a signature valid.
+VERIFY_TIMESTAMP_THRESHOLD: int = 1
 
 
 class Verifier:
@@ -81,8 +86,6 @@ class Verifier:
             for parent_cert in trusted_root.get_fulcio_certs()
         ]
         self._trusted_root = trusted_root
-
-        self.verify_timestamp_threshold: int = 1
 
     @classmethod
     def production(cls, *, offline: bool = False) -> Verifier:
@@ -118,7 +121,7 @@ class Verifier:
 
     def _verify_signed_timestamp(
         self, timestamp_response: TimeStampResponse, signature: bytes
-    ) -> bool:
+    ) -> TimestampVerificationResult | None:
         """
         Verify a Signed Timestamp using the TSA provided by the Trusted Root.
         """
@@ -127,9 +130,8 @@ class Verifier:
             certificates = certificate_authority.certificates(allow_expired=True)
 
             builder = VerifierBuilder()
-            for certificate in certificates[:-1]:
-                builder.add_intermediate_certificate(certificate)
-            builder.add_root_certificate(certificates[-1])
+            for certificate in certificates:
+                builder.add_root_certificate(certificate)
 
             verifier = builder.build()
             try:
@@ -148,7 +150,10 @@ class Verifier:
                     <= timestamp_response.tst_info.gen_time
                     < certificate_authority.validity_period_end
                 ):
-                    return True
+                    return TimestampVerificationResult(
+                        source=TimestampSource.TIMESTAMP_AUTHORITY,
+                        time=timestamp_response.tst_info.gen_time,
+                    )
 
                 _logger.debug(
                     "Unable to verify Timestamp because not in CA time range."
@@ -158,16 +163,18 @@ class Verifier:
                     "Unable to verify Timestamp because no validity provided."
                 )
 
-        return False
+        return None
 
-    def _verify_timestamp_authority(self, bundle: Bundle) -> int:
+    def _verify_timestamp_authority(
+        self, bundle: Bundle
+    ) -> List[TimestampVerificationResult]:
         """
         Verify that the given bundle has been timestamped by a trusted timestamp authority
         and that the timestamp is valid.
 
         Returns the number of valid signed timestamp in the bundle.
         """
-        timestamp_responses: List[TimeStampResponse] = (
+        timestamp_responses = (
             bundle.verification_material.timestamp_verification_data.rfc3161_timestamps
         )
         if len(timestamp_responses) > MAX_ALLOWED_TIMESTAMP:
@@ -181,10 +188,85 @@ class Verifier:
         # The Signer sends a hash of the signature as the messageImprint in a TimeStampReq
         # to the Timestamping Service
         signature_hash = sha256_digest(bundle.signature).digest
-        return [
-            self._verify_signed_timestamp(tsr, signature_hash)
-            for tsr in timestamp_responses
-        ].count(True)
+        verified_timestamps = []
+        for tsr in timestamp_responses:
+            if verified_timestamp := self._verify_signed_timestamp(tsr, signature_hash):
+                verified_timestamps.append(verified_timestamp)
+
+        return verified_timestamps
+
+    def _establish_time(self, bundle: Bundle) -> List[TimestampVerificationResult]:
+        """
+        Establish the time for bundle verification.
+
+        This method uses timestamps from two possible sources:
+        1. RFC3161 signed timestamps from a Timestamping Authority (TSA)
+        2. Transparency Log timestamps
+        """
+        verified_timestamps = []
+
+        # If a timestamp from the timestamping service is available, the Verifier MUST
+        # perform path validation using the timestamp from the Timestamping Service.
+        if bundle.verification_material.timestamp_verification_data.rfc3161_timestamps:
+            if not self._trusted_root.get_timestamp_authorities():
+                msg = (
+                    "no Timestamp Authorities have been provided to validate this "
+                    "bundle but it contains a signed timestamp"
+                )
+                raise VerificationError(msg)
+
+            timestamp_from_tsa = self._verify_timestamp_authority(bundle)
+            if len(timestamp_from_tsa) < VERIFY_TIMESTAMP_THRESHOLD:
+                msg = (
+                    f"not enough timestamps validated to meet the validation "
+                    f"threshold ({len(timestamp_from_tsa)}/{VERIFY_TIMESTAMP_THRESHOLD})"
+                )
+                raise VerificationError(msg)
+
+            verified_timestamps.extend(timestamp_from_tsa)
+
+        # If a timestamp from the Transparency Service is available, the Verifier MUST
+        # perform path validation using the timestamp from the Transparency Service.
+        if timestamp := bundle.log_entry.integrated_time:
+            verified_timestamps.append(
+                TimestampVerificationResult(
+                    source=TimestampSource.TRANSPARENCY_SERVICE,
+                    time=datetime.fromtimestamp(timestamp, tz=timezone.utc),
+                )
+            )
+        return verified_timestamps
+
+    def _verify_chain_at_time(
+        self, certificate: X509, timestamp_result: TimestampVerificationResult
+    ) -> List[X509]:
+        """
+        Verify the validity of the certificate chain at the given time.
+
+        Raises a VerificationError if the chain can't be built or be verified.
+        """
+        # NOTE: The `X509Store` object cannot have its time reset once the `set_time`
+        # method been called on it. To get around this, we construct a new one in each
+        # call.
+        store = X509Store()
+        # NOTE: By explicitly setting the flags here, we ensure that OpenSSL's
+        # PARTIAL_CHAIN default does not change on us. Enabling PARTIAL_CHAIN
+        # would be strictly more conformant of OpenSSL, but we currently
+        # *want* the "long" chain behavior of performing path validation
+        # down to a self-signed root.
+        store.set_flags(X509StoreFlags.X509_STRICT)
+        for parent_cert_ossl in self._fulcio_certificate_chain:
+            store.add_cert(parent_cert_ossl)
+
+        store.set_time(timestamp_result.time)
+
+        store_ctx = X509StoreContext(store, certificate)
+
+        try:
+            # get_verified_chain returns the full chain including the end-entity certificate
+            # and chain should contain only CA certificates
+            return store_ctx.get_verified_chain()[1:]
+        except X509StoreContextError as e:
+            raise VerificationError(f"failed to build chain: {e}")
 
     def _verify_common_signing_cert(
         self, bundle: Bundle, policy: VerificationPolicy
@@ -198,6 +280,7 @@ class Verifier:
 
         # In order to verify an artifact, we need to achieve the following:
         #
+        # 0. Establish a time for the signature.
         # 1. Verify that the signing certificate chains to the root of trust
         #    and is valid at the time of signing.
         # 2. Verify the signing certificate's SCT.
@@ -213,7 +296,7 @@ class Verifier:
         # 8. Verify the transparency log entry's consistency against the other
         #    materials, to prevent variants of CVE-2022-36056.
         #
-        # This method performs steps (1) through (6) above. Its caller
+        # This method performs steps (0) through (6) above. Its caller
         # MUST perform steps (7) and (8) separately, since they vary based on
         # the kind of verification being performed (i.e. hashedrekord, DSSE, etc.)
 
@@ -238,38 +321,17 @@ class Verifier:
         # While this step is optional and only performed if timestamp data has been
         # provided within the bundle, providing a signed timestamp without a TSA to
         # verify it result in a VerificationError.
-        if bundle.verification_material.timestamp_verification_data.rfc3161_timestamps:
-            if not self._trusted_root.get_timestamp_authorities():
-                msg = (
-                    "No Timestamp Authorities have been provided to validate this "
-                    "bundle but it contains a signed timestamp"
-                )
-                raise VerificationError(msg)
-
-            verified_timestamp = self._verify_timestamp_authority(bundle)
-            # The threshold is set to (1) by default but kept as a variable to allow
-            # this value to change
-            if verified_timestamp < self.verify_timestamp_threshold:
-                msg = (
-                    f"Not enough Timestamp validated to meet the Validation "
-                    f"Threshold ({verified_timestamp}/{self.verify_timestamp_threshold})"
-                )
-                raise VerificationError(msg)
+        verified_timestamps = self._establish_time(bundle)
+        if not verified_timestamps:
+            raise VerificationError("not enough sources of verified time")
 
         # (1): verify that the signing certificate is signed by the root
         #      certificate and that the signing certificate was valid at the
         #      time of signing.
-        sign_date = cert.not_valid_before_utc
         cert_ossl = X509.from_cryptography(cert)
-
-        store.set_time(sign_date)
-        store_ctx = X509StoreContext(store, cert_ossl)
-        try:
-            # get_verified_chain returns the full chain including the end-entity certificate
-            # and chain should contain only CA certificates
-            chain = store_ctx.get_verified_chain()[1:]
-        except X509StoreContextError as e:
-            raise VerificationError(f"failed to build chain: {e}")
+        chain: list[X509] = []
+        for vts in verified_timestamps:
+            chain = self._verify_chain_at_time(cert_ossl, vts)
 
         # (2): verify the signing certificate's SCT.
         sct = _get_precertificate_signed_certificate_timestamps(cert)[0]